fix(cartera-back): CB-027 — email_asesor usa emailCashIn, test sin DB
- email_asesor filtraba platform_users.email, campo documentado como
  desactualizado para varios asesores (poolPorAsesor.ts) mientras
  asesores.emailCashIn es el que coincide con el email real del CRM — y el
  mismo que este endpoint ya devuelve como asesor_email. Filtrar por el
  campo mostrado, no por otro que puede divergir.
- calcularProgresoConvenio se mueve a paymentAgreement-helpers.ts (sin
  imports de DB) para que su test no arrastre database/index.ts, que lanza
  en import-time si SUPABASE_DB_URL no está configurado.

Addresses review comments from chatgpt-codex-connector on PR #1213.
